Reworked the transformers and enabled the implementation to run the old code but adapted.
