+++ src/main/java/net/fabricmc/loom/util/ModProcessor.java
@@ -33,6 +33,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.function.Function;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 import java.util.zip.ZipEntry;
@@ -58,74 +59,68 @@ import net.fabricmc.tinyremapper.TinyRemapper;
 public class ModProcessor {
 	private static final Gson GSON = new Gson();
 
-	public static void processMod(File input, File output, Project project, Configuration config, ResolvedArtifact artifact) throws IOException {
-		if (output.exists()) {
-			output.delete();
-		}
-
-		remapJar(input, output, project, artifact);
+	public static void processMod(Project project, File input, Function<String, File> outputFileProducer) throws IOException {
+		final File output = outputFileProducer.apply(input.getName());
+	    remapJar(input, output, outputFileProducer, project);
 
 		//Enable this if you want your nested jars to be extracted, this will extract **all** jars
 		if (project.getExtensions().getByType(LoomGradleExtension.class).extractJars) {
-			handleNestedJars(input, project, config, artifact);
+			handleNestedJars(project, input, outputFileProducer);
 		}
 
 		//Always strip the nested jars
 		stripNestedJars(output);
 	}
 
-	private static void handleNestedJars(File input, Project project, Configuration config, ResolvedArtifact artifact) throws IOException {
-		JarFile jarFile = new JarFile(input);
-		JarEntry modJsonEntry = jarFile.getJarEntry("fabric.mod.json");
+    private static void handleNestedJars(
+                    Project project,
+                    File input,
+                    Function<String, File> outputFileProducer) throws IOException {
+        JarFile jarFile = new JarFile(input);
+        JarEntry modJsonEntry = jarFile.getJarEntry("fabric.mod.json");
 
-		if (modJsonEntry == null) {
-			return;
-		}
+        if (modJsonEntry == null) {
+            return;
+        }
 
-		try (InputStream inputStream = jarFile.getInputStream(modJsonEntry)) {
-			JsonObject json = GSON.fromJson(new InputStreamReader(inputStream), JsonObject.class);
+        try (InputStream inputStream = jarFile.getInputStream(modJsonEntry)) {
+            JsonObject json = GSON.fromJson(new InputStreamReader(inputStream), JsonObject.class);
 
-			if (json == null || !json.has("jars")) {
-				return;
-			}
+            if (json == null || !json.has("jars")) {
+                return;
+            }
 
-			JsonArray jsonArray = json.getAsJsonArray("jars");
+            JsonArray jsonArray = json.getAsJsonArray("jars");
 
-			for (int i = 0; i < jsonArray.size(); i++) {
-				JsonObject jsonObject = jsonArray.get(i).getAsJsonObject();
-				String fileName = jsonObject.get("file").getAsString();
-				project.getLogger().lifecycle(String.format("Found %s nested in %s", fileName, input.getName()));
-				processNestedJar(jarFile, fileName, project, config, artifact);
-			}
-		}
-	}
+            for (int i = 0; i < jsonArray.size(); i++) {
+                JsonObject jsonObject = jsonArray.get(i).getAsJsonObject();
+                String fileName = jsonObject.get("file").getAsString();
+                project.getLogger().lifecycle(String.format("Found %s nested in %s", fileName, input.getName()));
+                processNestedJar(project, jarFile, fileName, outputFileProducer);
+            }
+        }
+    }
 
-	private static void processNestedJar(JarFile parentJar, String fileName, Project project, Configuration config, ResolvedArtifact artifact) throws IOException {
-		LoomGradleExtension extension = project.getExtensions().getByType(LoomGradleExtension.class);
+    private static void processNestedJar(
+                    Project project,
+                    JarFile parentJar,
+                    String fileName,
+                    Function<String, File> outputFileProducer) throws IOException {
+        JarEntry entry = parentJar.getJarEntry(fileName);
 
-		JarEntry entry = parentJar.getJarEntry(fileName);
+        if (entry == null) {
+            throw new RuntimeException(String.format("%s was not found in %s", fileName, parentJar.getName()));
+        }
 
-		if (entry == null) {
-			throw new RuntimeException(String.format("%s was not found in %s", fileName, parentJar.getName()));
-		}
+        File nestedFile = File.createTempFile(String.format("nested_%s", fileName), "jar");
+        nestedFile.deleteOnExit();
 
-		File nestedFile = new File(extension.getNestedModCache(), fileName.substring(fileName.lastIndexOf("/")));
+        try (InputStream jarStream = parentJar.getInputStream(entry)) {
+            FileUtils.copy(jarStream, nestedFile);
+        }
 
-		try (InputStream jarStream = parentJar.getInputStream(entry)) {
-			FileUtils.copy(jarStream, nestedFile);
-		}
-
-		File remappedFile = new File(extension.getRemappedModCache(), fileName.substring(fileName.lastIndexOf("/")));
-
-		processMod(nestedFile, remappedFile, project, config, artifact);
-
-		if (!remappedFile.exists()) {
-			throw new RuntimeException("Failed to find processed nested jar");
-		}
-
-		//Add the project right onto the remapped mods, hopefully this works
-		project.getDependencies().add(config.getName(), project.files(remappedFile));
-	}
+        processMod(project, nestedFile, outputFileProducer);
+    }
 
 	private static void stripNestedJars(File file) {
 		//Strip out all contained jar info as we dont want loader to try and load the jars contained in dev.
@@ -139,7 +134,7 @@ public class ModProcessor {
 		}))});
 	}
 
-	private static void remapJar(File input, File output, Project project, ResolvedArtifact artifact) throws IOException {
+	private static void remapJar(File input, File output, Function<String, File> outputFileProducer, Project project) throws IOException {
 		LoomGradleExtension extension = project.getExtensions().getByType(LoomGradleExtension.class);
 		String fromM = "intermediary";
 		String toM = "named";
@@ -164,14 +159,8 @@ public class ModProcessor {
 
 		project.getLogger().lifecycle(":remapping " + input.getName() + " (TinyRemapper, " + fromM + " -> " + toM + ")");
 
-		// If the sources don't exist, we want remapper to give nicer names to the missing variable names.
-		// However, if the sources do exist, if remapper gives names to the parameters that prevents IDEs (at least IDEA)
-		// from replacing the parameters with the actual names from the sources.
-		boolean sourcesExist = ModCompileRemapper.findSources(project.getDependencies(), artifact) != null;
-
 		TinyRemapper remapper = TinyRemapper.newRemapper()
 						.withMappings(TinyRemapperMappingsHelper.create(mappingsProvider.getMappings(), fromM, toM, false))
-						.renameInvalidLocals(!sourcesExist)
 						.build();
 
 		try (OutputConsumerPath outputConsumer = new OutputConsumerPath.Builder(Paths.get(output.getAbsolutePath())).build()) {
@@ -188,47 +177,5 @@ public class ModProcessor {
 		if (!output.exists()) {
 			throw new RuntimeException("Failed to remap JAR to " + toM + " file not found: " + output.getAbsolutePath());
 		}
-	}
-
-	static JsonObject readInstallerJson(File file, Project project) {
-		try {
-			LoomGradleExtension extension = project.getExtensions().getByType(LoomGradleExtension.class);
-			String launchMethod = extension.getLoaderLaunchMethod();
-
-			String jsonStr;
-			int priority = 0;
-
-			try (JarFile jarFile = new JarFile(file)) {
-				ZipEntry entry = null;
-
-				if (!launchMethod.isEmpty()) {
-					entry = jarFile.getEntry("fabric-installer." + launchMethod + ".json");
-
-					if (entry == null) {
-						project.getLogger().warn("Could not find loader launch method '" + launchMethod + "', falling back");
-					}
-				}
-
-				if (entry == null) {
-					entry = jarFile.getEntry("fabric-installer.json");
-					priority++;
-
-					if (entry == null) {
-						return null;
-					}
-				}
-
-				try (InputStream inputstream = jarFile.getInputStream(entry)) {
-					jsonStr = IOUtils.toString(inputstream, StandardCharsets.UTF_8);
-				}
-			}
-
-			JsonObject jsonObject = GSON.fromJson(jsonStr, JsonObject.class);
-			return jsonObject;
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-
-		return null;
 	}
 }
